code/mlflow-intro/pyproject.toml
[tool.poetry]
name = "mlflow-intro"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
mlflow = "^2.11.1"
notebook = "^7.1.1"
matplotlib = "^3.8.3"
numpy = "^1.26.4"
scipy = "^1.12.0"
scikit-learn = "^1.4.1.post1"
seaborn = "^0.13.2"
tensorflow-datasets = "^4.9.4"
pillow = "^10.2.0"


[tool.poetry.group.tf_mac]
optional = true

[tool.poetry.group.tf_mac.dependencies]
tensorflow-macos = "^2.13.0"
tensorflow-metal = "^1.0.1"

[tool.poetry.group.tf_win]
optional = true

[tool.poetry.group.tf_win.dependencies]
tensorflow = "^2.13.0"

[tool.poetry.group.tf_linux]
optional = true

[tool.poetry.group.tf_linux.dependencies]
tensorflow = "^2.13.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
